+++ skbio/core/alignment/__init__.py
@@ -29,7 +29,7 @@ Alignment Algorithms
 
    StripedSmithWaterman
    AlignmentStructure
-   align_striped_smith_waterman
+   local_pairwise_align_ssw
 
 Data Structure Examples
 -----------------------
@@ -74,26 +74,19 @@ seq2          TCC--G-GGGA
 
 Alignment Algorithm Examples
 ----------------------------
-Using the convenient ``align_striped_smith_waterman`` function:
+Using the convenient ``local_pairwise_align_ssw`` function:
 
->>> from skbio.core.alignment import align_striped_smith_waterman
->>> alignment = align_striped_smith_waterman(
+>>> from skbio.core.alignment import local_pairwise_align_ssw
+>>> alignment = local_pairwise_align_ssw(
 ...                 "ACTAAGGCTCTCTACCCCTCTCAGAGA",
 ...                 "ACTAAGGCTCCTAACCCCCTTTTCTCAGA"
 ...             )
 >>> print alignment
-{
-    'optimal_alignment_score': 27,
-    'suboptimal_alignment_score': 21,
-    'query_begin': 0,
-    'query_end': 24,
-    'target_begin': 0,
-    'target_end_optimal': 28,
-    'target_end_suboptimal': 12,
-    'cigar': '10M1I2M1D5M4D7M',
-    'query_sequence': 'ACTAAGGCTCTCTACCCCTCTCAGAGA',
-    'target_sequence': 'ACTAAGGCTCCTAACCCCCTTTTCTCAGA'
-}
+>query
+ACTAAGGCTCTC-TACCC----CTCTCAGA
+>target
+ACTAAGGCTC-CTAACCCCCTTTTCTCAGA
+<BLANKLINE>
 
 Using the ``StripedSmithWaterman`` object:
 
@@ -101,18 +94,10 @@ Using the ``StripedSmithWaterman`` object:
 >>> query = StripedSmithWaterman("ACTAAGGCTCTCTACCCCTCTCAGAGA")
 >>> alignment = query("AAAAAACTCTCTAAACTCACTAAGGCTCTCTACCCCTCTTCAGAGAAGTCGA")
 >>> print alignment
-{
-    'optimal_alignment_score': 49,
-    'suboptimal_alignment_score': 24,
-    'query_begin': 0,
-    'query_end': 26,
-    'target_begin': 18,
-    'target_end_optimal': 45,
-    'target_end_suboptimal': 29,
-    'cigar': '20M1D7M',
-    'query_sequence': 'ACTAAGGCTCTCTACCCCTCTCAGAGA',
-    'target_sequence': 'AAAAAACTCTCTAAACTCACTAAGGCTCTCTACCCCTCTTCAGAGAAGTCGA'
-}
+ACTAAGGCTC...
+ACTAAGGCTC...
+Score: 49
+Length: 28
 
 Using the ``StripedSmithWaterman`` object for multiple targets in an efficient
 way and finding the aligned sequence representations:
@@ -132,21 +117,13 @@ way and finding the aligned sequence representations:
 ...     alignments.append(alignment)
 ...
 >>> print alignments[0]
-{
-    'optimal_alignment_score': 38,
-    'suboptimal_alignment_score': 14,
-    'query_begin': 0,
-    'query_end': 26,
-    'target_begin': 4,
-    'target_end_optimal': 32,
-    'target_end_suboptimal': 15,
-    'cigar': '3M1I6M3D17M',
-    'query_sequence': 'ACTAAGGCTCTCTACCCCTCTCAGAGA',
-    'target_sequence': 'GCTAACTAGGCTCCCTTCTACCCCTCTCAGAGA'
-}
->>> print alignments[0].get_aligned_query_sequence()
+ACTAAGGCT-...
+ACT-AGGCTC...
+Score: 38
+Length: 30
+>>> print alignments[0].aligned_query_sequence
 ACTAAGGCT---CTCTACCCCTCTCAGAGA
->>> print alignments[0].get_aligned_target_sequence()
+>>> print alignments[0].aligned_target_sequence
 ACT-AGGCTCCCTTCTACCCCTCTCAGAGA
 
 """
@@ -161,11 +138,11 @@ ACT-AGGCTCCCTTCTACCCCTCTCAGAGA
 
 from .alignment import Alignment, SequenceCollection, StockholmAlignment
 from .ssw.ssw_wrapper import (
-    StripedSmithWaterman, AlignmentStructure, align_striped_smith_waterman)
+    StripedSmithWaterman, local_pairwise_align_ssw, AlignmentStructure)
 
 __all__ = ['Alignment', 'SequenceCollection', 'StockholmAlignment',
            'StripedSmithWaterman', 'AlignmentStructure',
-           'align_striped_smith_waterman']
+           'local_pairwise_align_ssw']
 
 from numpy.testing import Tester
 test = Tester().test